requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
pyyaml>=6.0
onnxruntime>=1.16.0

# Optional: for Apple Silicon optimization (CoreML provider)
# Install with: pip install onnxruntime-silicon>=1.16.0
